pyproject.toml
[tool.poetry]
name = "hotels-app"
version = "0.1.0"
description = ""
authors = ["Artur Mistiuk <mistiuk.artur@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.104.0"}
sqlalchemy = "^2.0.23"
alembic = "^1.12.1"
asyncpg = "^0.29.0"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "3.2.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.2.1"}
celery = "^5.3.6"
flower = "^2.0.1"
pillow = "^10.1.0"
sqladmin = "^0.16.0"
pytest = "^7.4.4"
httpx = "^0.26.0"
pytest-asyncio = "^0.23.3"
flake8 = "^7.0.0"
isort = "^5.13.2"
python-json-logger = "^2.0.7"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
